+++ CHANGELOG.md
@@ -4,10 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased] - 2022-05-07
+## [Unreleased] - 2022-05-09
 
 ### Added
 * Streams of binomial distributed integers.
+* Streams of not strictly uniform random integers.
 * Seeded constructors for the EnhancedSplittableGenerator, EnhancedStreamableGenerator, and
   EnhancedRandomGenerator classes.